Do not require icu unless you really need it.
