+++ searchcore/src/apps/proton/CMakeLists.txt
@@ -26,4 +26,5 @@ vespa_add_executable(searchcore_proton_app
     searchcore_fconfig
     searchcore_util
     storageserver_storageapp
+    searchlib_searchlib_uca
 )
