+++ searchcore/src/apps/tests/CMakeLists.txt
@@ -22,4 +22,5 @@ vespa_add_executable(searchcore_persistenceconformance_test_app TEST
     searchcore_util
     vdstestlib
     persistence_persistence_conformancetest
+    searchlib_searchlib_uca
 )
