+++ searchcore/src/tests/proton/docsummary/CMakeLists.txt
@@ -22,6 +22,7 @@ vespa_add_executable(searchcore_docsummary_test_app TEST
     searchcore_proton_metrics
     searchcore_fconfig
     searchcore_util
+    searchlib_searchlib_uca
 )
 vespa_add_executable(searchcore_summaryfieldconverter_test_app
     SOURCES
