+++ searchcore/src/tests/proton/documentdb/CMakeLists.txt
@@ -20,6 +20,7 @@ vespa_add_executable(searchcore_documentdb_test_app TEST
     searchcore_proton_metrics
     searchcore_fconfig
     searchcore_util
+    searchlib_searchlib_uca
 )
 vespa_add_test(NAME searchcore_documentdb_test_app COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/documentdb_test.sh
                DEPENDS searchcore_documentdb_test_app)
