+++ searchlib/src/vespa/searchlib/CMakeLists.txt
@@ -32,6 +32,5 @@ vespa_add_library(searchlib
     INSTALL lib64
     DEPENDS
     staging_vespalib
-    icuuc
     atomic
 )
